pyproject.toml
[project]
name = "bria_client"
dynamic = ["version"]
description = "Bria.ai Python SDK"
readme = "README.md"
authors = [{ name = "Bria.ai" }]
maintainers = [
    { name = "omri", email = "omriassa@bria.ai" },
    { name = "noam", email = "noam@bria.ai" }
]

requires-python = ">=3.10"
dependencies = [
    "httpx~=0.28",
    "httpx-retries~=0.4.5",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "strenum~=0.4.15; python_version < '3.11'",
    "werkzeug>=3",
]

[tool.uv-dynamic-versioning]
pattern = "default-unprefixed"
fallback-version = "0.0.0"

[tool.uv]
cache-keys = [{ file = "pyproject.toml" }, { git = { commit = true, tags = true } }]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bria_client"]

[tool.hatch.build]
include = ["src/**/*"]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.ruff]
line-length = 160
target-version = "py310"

[tool.ruff.lint]
select = ["C4", "E", "F", "I", "PERF", "UP"]
ignore = ["PERF203", "E402", "F821"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pyright>=1.1.407",
    "ruff>=0.14.0",
]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
markers = [
    "unit: fast unit tests",
    "integration: tests requiring external resources",
    "e2e: full system end-to-end tests",
]
testpaths = ["tests"]
log_cli = true

[tool.pyright]
enableExperimentalFeatures = true
